fixing go install

+++ scripts/goland.sh
@@ -1,8 +1,10 @@
 #!/usr/bin/env bash
 
-wget https://download.jetbrains.com/go/goland-2018.3.3.tar.gz 
-sudo tar -C "/usr/sbin/" -xzf goland-2018.3.3.tar.gz
-sudo ln -s /usr/sbin/GoLand-2018.3.3/bin/goland.sh /usr/bin/goland
-rm goland-2018.3.3.tar.gz
+if [ ! -f goland-2018.3.3.tar.gz ]; then
+	wget https://download.jetbrains.com/go/goland-2018.3.3.tar.gz 
+	sudo tar -C "/usr/sbin/" -xzf goland-2018.3.3.tar.gz
+	sudo ln -s /usr/sbin/GoLand-2018.3.3/bin/goland.sh /usr/bin/goland
+fi
+
